refactor: Move Remote config initialization to Application class
