+++ app/src/main/java/campus/tech/kakao/map/ui/SplashActivity.kt
@@ -2,32 +2,41 @@ package campus.tech.kakao.map.ui
 
 import android.content.Intent
 import android.os.Bundle
-import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
 import androidx.databinding.DataBindingUtil
+import campus.tech.kakao.map.KakaoMapApplication
 import campus.tech.kakao.map.R
 import campus.tech.kakao.map.databinding.SplashLayoutBinding
 import dagger.hilt.android.AndroidEntryPoint
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.delay
+import kotlinx.coroutines.launch
 
 @AndroidEntryPoint
 class SplashActivity : AppCompatActivity() {
 
     private lateinit var splashBinding: SplashLayoutBinding
-    private val splashViewModel: SplashViewModel by viewModels()
+    private val splashDisplayLength: Long = 1000
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         splashBinding = DataBindingUtil.setContentView(this, R.layout.splash_layout)
 
-        splashViewModel.serviceState.observe(this) { serviceState ->
-            if (serviceState == "ON_SERVICE") {
-                val intent = Intent(this, MapActivity::class.java)
-                startActivity(intent)
-            }
-        }
+        val remoteConfig = KakaoMapApplication.remoteConfig
+        val serviceState = remoteConfig.getString("serviceState")
+        val serviceMessage = remoteConfig.getString("serviceMessage")
 
-        splashViewModel.serviceMessage.observe(this) { serviceMessage ->
-            splashBinding.tvError.text = serviceMessage
+        CoroutineScope(Dispatchers.Main).launch {
+            delay(splashDisplayLength)
+            if (serviceState == "ON_SERVICE") {
+                val intent = Intent(this@SplashActivity, MapActivity::class.java)
+                startActivity(intent)
+                finish()
+            }
+            else {
+                splashBinding.tvError.text = serviceMessage
+            }
         }
     }
 }
