--- app/src/main/java/campus/tech/kakao/map/ui/SplashViewModel.kt
@@ -1,34 +1,0 @@
-package campus.tech.kakao.map.ui
-
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-import com.google.firebase.remoteconfig.FirebaseRemoteConfig
-import dagger.hilt.android.lifecycle.HiltViewModel
-import javax.inject.Inject
-
-@HiltViewModel
-class SplashViewModel @Inject constructor(private val remoteConfig: FirebaseRemoteConfig) : ViewModel() {
-    private val _serviceState = MutableLiveData<String>()
-    val serviceState: LiveData<String> get() = _serviceState
-
-    private val _serviceMessage = MutableLiveData<String>()
-    val serviceMessage: LiveData<String> get() = _serviceMessage
-
-    init {
-        fetchRemoteConfig()
-    }
-
-    private fun fetchRemoteConfig() {
-        remoteConfig.fetchAndActivate().addOnCompleteListener { task ->
-            if (task.isSuccessful) {
-                val serviceState = remoteConfig.getString("serviceState")
-                _serviceState.value = serviceState
-            }
-            else {
-                val serviceMessage = remoteConfig.getString("serviceMessage")
-                _serviceMessage.value = serviceMessage
-            }
-        }
-    }
-}
