+++ app/src/main/res/values/strings.xml
@@ -3,7 +3,5 @@
     <string name="search_placeholder">검색어를 입력해 주세요.</string>
     <string name="search_no_data">검색 결과가 없습니다.</string>
     <string name="BASE_URL">https://dapi.kakao.com</string>
-    <string name="msg_token_fmt">FCM Registration Token: %1$s</string>
-    <string name="channel_name">main channelName</string>
     <string name="channel_description">notification</string>
 </resources>
